[libunwind][Cygwin] Add support for Cygwin target

+++ libunwind/src/Unwind-seh.cpp
@@ -93,7 +93,7 @@ _GCC_specific_handler(PEXCEPTION_RECORD ms_exc, PVOID frame, PCONTEXT ms_ctx,
   bool ours = false;
 
   _LIBUNWIND_TRACE_UNWINDING("_GCC_specific_handler(%#010lx(%lx), %p)",
-                             ms_exc->ExceptionCode, ms_exc->ExceptionFlags,
+                             (unsigned long)ms_exc->ExceptionCode, (unsigned long)ms_exc->ExceptionFlags,
                              (void *)frame);
   if (ms_exc->ExceptionCode == STATUS_GCC_UNWIND) {
     if (IS_TARGET_UNWIND(ms_exc->ExceptionFlags)) {
@@ -148,7 +148,7 @@ _GCC_specific_handler(PEXCEPTION_RECORD ms_exc, PVOID frame, PCONTEXT ms_ctx,
 
   _LIBUNWIND_TRACE_UNWINDING("_GCC_specific_handler() calling personality "
                              "function %p(1, %d, %llx, %p, %p)",
-                             (void *)pers, action, exc->exception_class,
+                             (void *)pers, action, (unsigned long long)exc->exception_class,
                              (void *)exc, (void *)ctx);
   urc = pers(1, action, exc->exception_class, exc, ctx);
   _LIBUNWIND_TRACE_UNWINDING("_GCC_specific_handler() personality returned %d", urc);
